new cli skeleton - wip
