+++ cli/commands/root.go
@@ -1,0 +1,17 @@
+package commands
+
+import (
+	"github.com/spf13/cobra"
+	"os"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "bcdbadmin",
+	Short: "Config Orion via CLI.",
+}
+
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		os.Exit(1)
+	}
+}
